chore: add permission namespace
