--- packages/sheets/src/facade/const.ts
@@ -1,62 +1,0 @@
-/**
- * Copyright 2023-present DreamNum Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { RangeProtectionPermissionEditPoint, RangeProtectionPermissionViewPoint, WorkbookCommentPermission, WorkbookCopyPermission, WorkbookCreateProtectPermission, WorkbookCreateSheetPermission, WorkbookDeleteSheetPermission, WorkbookDuplicatePermission, WorkbookEditablePermission, WorkbookExportPermission, WorkbookHideSheetPermission, WorkbookHistoryPermission, WorkbookManageCollaboratorPermission, WorkbookMoveSheetPermission, WorkbookPrintPermission, WorkbookRecoverHistoryPermission, WorkbookRenameSheetPermission, WorkbookSharePermission, WorkbookViewHistoryPermission, WorkbookViewPermission, WorksheetCopyPermission, WorksheetDeleteColumnPermission, WorksheetDeleteProtectionPermission, WorksheetDeleteRowPermission, WorksheetEditExtraObjectPermission, WorksheetEditPermission, WorksheetFilterPermission, WorksheetInsertColumnPermission, WorksheetInsertHyperlinkPermission, WorksheetInsertRowPermission, WorksheetManageCollaboratorPermission, WorksheetPivotTablePermission, WorksheetSelectProtectedCellsPermission, WorksheetSelectUnProtectedCellsPermission, WorksheetSetCellStylePermission, WorksheetSetCellValuePermission, WorksheetSetColumnStylePermission, WorksheetSetRowStylePermission, WorksheetSortPermission, WorksheetViewPermission } from '@univerjs/sheets';
-
-export const PermissionPointsDefinitions = {
-    WorkbookCommentPermission,
-    WorkbookCopyPermission,
-    WorkbookCreateProtectPermission,
-    WorkbookCreateSheetPermission,
-    WorkbookDeleteSheetPermission,
-    WorkbookDuplicatePermission,
-    WorkbookEditablePermission,
-    WorkbookExportPermission,
-    WorkbookHideSheetPermission,
-    WorkbookHistoryPermission,
-    WorkbookManageCollaboratorPermission,
-    WorkbookMoveSheetPermission,
-    WorkbookPrintPermission,
-    WorkbookRecoverHistoryPermission,
-    WorkbookRenameSheetPermission,
-    WorkbookSharePermission,
-    WorkbookViewHistoryPermission,
-    WorkbookViewPermission,
-    WorksheetCopyPermission,
-
-    WorksheetDeleteColumnPermission,
-    WorksheetDeleteProtectionPermission,
-    WorksheetDeleteRowPermission,
-    WorksheetEditExtraObjectPermission,
-    WorksheetEditPermission,
-    WorksheetFilterPermission,
-    WorksheetInsertColumnPermission,
-    WorksheetInsertHyperlinkPermission,
-    WorksheetInsertRowPermission,
-    WorksheetManageCollaboratorPermission,
-    WorksheetPivotTablePermission,
-    WorksheetSelectProtectedCellsPermission,
-    WorksheetSelectUnProtectedCellsPermission,
-    WorksheetSetCellStylePermission,
-    WorksheetSetCellValuePermission,
-    WorksheetSetColumnStylePermission,
-    WorksheetSetRowStylePermission,
-    WorksheetSortPermission,
-    WorksheetViewPermission,
-
-    RangeProtectionPermissionEditPoint,
-    RangeProtectionPermissionViewPoint,
-};
